ADD: city attractions and town halls can be defined to be build for more than 65535 citizens
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@6094 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/haus_besch.h
@@ -175,7 +175,7 @@ class haus_besch_t : public obj_besch_std_name_t { // Daten für ein ganzes Gebäu
 	utyp            utype; // Hajo: if gtyp == gebaeude_t::unbekannt, then this is the real type
 
 	uint16 animation_time;	// in ms
-	uint16 extra_data;      // bauzeit (inhabitants) for city attractions, waytype for depots, player level for headquarters
+	uint32 extra_data;      // bauzeit (inhabitants) for city attractions, waytype for depots, player level for headquarters
 	koord  groesse;
 	flag_t flags;
 	uint16 level;          // or passengers;
@@ -221,7 +221,7 @@ public:
 
 	uint8 get_all_layouts() const { return layouts; }
 
-	uint16 get_extra() const { return extra_data; }
+	uint32 get_extra() const { return extra_data; }
 
 	// ground is transparent
 	bool ist_mit_boden() const { return (flags & FLAG_NEED_GROUND) != 0; }
